Update SuperLinter EditorConfig Checker (#4)
# Pull Request

## Description

This pull request includes updates to the `.editorconfig-checker`
configuration and the GitHub Actions workflow for code checks. The
changes address a typo in a configuration key, rename a configuration
file, and update the workflow to reflect the new file name.

### Updates to `.editorconfig-checker` configuration:

* Fixed a typo in the key `SpacesAftertabs`, correcting it to
`SpacesAfterTabs` in the `.editorconfig-checker` configuration file.
* Renamed the configuration file from `.ecrc` to
`.editorconfig-checker.json` for clarity and consistency.

### Updates to GitHub Actions workflow:

* Updated the `code-checks.yml` workflow to include a new environment
variable, `EDITORCONFIG_FILE_NAME`, pointing to the renamed
`.editorconfig-checker.json` file.