Swap SuperLinter to Full Version (#26)
# Pull Request

## Description

This pull request includes updates to dependencies and configurations
for code quality and pre-commit hooks. The most important changes
involve updating the version of `super-linter` in the GitHub Actions
workflow and increasing the minimum version requirement for `lefthook`.

### Updates to dependencies and configurations:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L27-L36):
Updated the `super-linter` version in the GitHub Actions workflow to a
specific commit hash referencing v7.4.0, replacing the previous slim
version.

*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Increased the `min_version` requirement for `lefthook` from 1.11.2 to
1.11.13.